compliation done

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -8,7 +8,7 @@
  *
  * Return: defined
  */
-int set_bit(unsigned long int *n, unsigned long int index)
+int set_bit(unsigned long int *n, unsigned int index)
 {
 	int addition;
